backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==1.1.0
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.0
marshmallow<4.0.0